+++ src/cone.c
@@ -9,7 +9,7 @@
 
 int calc_cone(int *t)
 {
-    int *equ = compute_quad_eq_cn(t);
+    float *equ = compute_quad_eq_cn(t);
 
     printf("Cone with a %d degree angle\n", t[7]);
     printf("Line passing through the point (%d, %d, %d) ", t[1], t[2], t[3]);
